Tested SSH, MYSQL

+++ start.sh
@@ -1,14 +1,11 @@
 #!/bin/bash
-if [ ! -f /vietcli-pw.txt ]; then
+if [ ! -f /home/vietcli/.log/vietcli-pw.txt ]; then
     #mysql has to be started this way as it doesn't work to call from /etc/init.d
     /usr/bin/mysqld_safe &
     sleep 10s
     # Here we generate random passwords (thank you pwgen!). The first two are for mysql users, the last batch for random keys in wp-config.php
     ROOT_PASSWORD=`pwgen -c -n -1 12`
     VIETCLI_PASSWORD="vietcli"
-
-    MYSQL_ROOT_PASSWORD=`pwgen -c -n -1 12`
-    MYSQL_VIETCLI_PASSWORD="vietcli"
 
     # echo "vietcli:$MAGENTO_PASSWORD" | chpasswd
     echo "root:$ROOT_PASSWORD" | chpasswd
@@ -17,25 +14,40 @@ if [ ! -f /vietcli-pw.txt ]; then
     mkdir /home/vietcli/.log
     echo root password: $ROOT_PASSWORD
     echo vietcli password: $VIETCLI_PASSWORD
-    echo mysql root password: $MYSQL_ROOT_PASSWORD
-    echo mysql vietcli password: $MYSQL_VIETCLI_PASSWORD
 
     echo $ROOT_PASSWORD > /home/vietcli/.log/root-pw.txt
     echo $VIETCLI_PASSWORD > /home/vietcli/.log/vietcli-pw.txt
-    echo $MYSQL_ROOT_PASSWORD > /mysql-vietcli-root-pw.txt
-    echo $MYSQL_VIETCLI_PASSWORD > /mysql-vietcli-pw.txt
-
-    mysqladmin -u root password $MYSQL_ROOT_PASSWORD
-    mysql -uroot -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY '$MYSQL_ROOT_PASSWORD' WITH GRANT OPTION; FLUSH PRIVILEGES;"
-    mysql -uroot -p$MYSQL_ROOT_PASSWORD -e "CREATE DATABASE vietcli_db; GRANT ALL PRIVILEGES ON vietcli_db.* TO 'vietcli'@'localhost' IDENTIFIED BY '$MYSQL_VIETCLI_PASSWORD'; FLUSH PRIVILEGES;"
-    killall mysqld
-    mv /var/lib/mysql/ibdata1 /var/lib/mysql/ibdata1.bak
-    cp -a /var/lib/mysql/ibdata1.bak /var/lib/mysql/ibdata1
 
     # Enable Magento 2 site
     ln -s /etc/nginx/sites-available/magento2.conf /etc/nginx/sites-enabled/
 
 fi
+
+
+if [ ! -f /home/vietcli/.log/mysql-vietcli-root-pw.txt ]; then
+    #mysql has to be started this way as it doesn't work to call from /etc/init.d
+    /usr/bin/mysqld_safe &
+    sleep 10s
+    # Here we generate random passwords (thank you pwgen!). The first two are for mysql users, the last batch for random keys in wp-config.php
+
+    MYSQL_ROOT_PASSWORD=`pwgen -c -n -1 12`
+    MYSQL_VIETCLI_PASSWORD="vietcli"
+
+    #This is so the passwords show up in logs.
+    echo mysql root password: $MYSQL_ROOT_PASSWORD
+    echo mysql vietcli password: $MYSQL_VIETCLI_PASSWORD
+
+    echo $MYSQL_ROOT_PASSWORD > /home/vietcli/.log/mysql-vietcli-root-pw.txt
+    echo $MYSQL_VIETCLI_PASSWORD > /home/vietcli/.log/mysql-vietcli-pw.txt
+
+    mysqladmin -u root -pvietcli password $MYSQL_ROOT_PASSWORD
+    mysql -uroot -p$MYSQL_ROOT_PASSWORD -e "GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY '$MYSQL_ROOT_PASSWORD' WITH GRANT OPTION; FLUSH PRIVILEGES;"
+    mysql -uroot -p$MYSQL_ROOT_PASSWORD -e "CREATE DATABASE vietcli_db; GRANT ALL PRIVILEGES ON vietcli_db.* TO 'vietcli'@'localhost' IDENTIFIED BY '$MYSQL_VIETCLI_PASSWORD'; FLUSH PRIVILEGES;"
+    killall mysqld
+
+    service mysql restart
+fi
+
 
 # Check HTTP_SERVER_NAME environment variable to set Virtual Host Name
 if [ -z "$HTTP_SERVER_NAME" ]; then
